requirements.txt
asgiref==3.2.7
astroid==2.4.1
autopep8==1.5.2
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.6
django-allauth==0.41.0
django-cors-headers==3.2.1
django-guardian==2.2.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
future==0.18.2
gunicorn==20.0.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
Pillow==7.1.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycodestyle==2.6.0
PyInvoice==0.1.7
PyJWT==1.7.1
pylint==2.5.2
python-http-client==3.2.7
python3-openid==3.1.0
pytz==2020.1
reportlab==3.5.42
requests==2.23.0
requests-oauthlib==1.3.0
sendgrid==6.3.1
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
whitenoise==5.0.1
wrapt==1.12.1
